If already logged in, redirect to root Remove jwt if get() failed
